feat: Add position.py file
The script extracts all unique positions from the players.csv file. The duplicates are dropped and
the rows with missing or invalid values are removed. The index is reset and a new column is added which
represents the position_id in the resulting Oracle database. The contents are written to a new file
called 'positions.csv'.